handle demo mode, avoid calling backend to approve payment if on demo

+++ backend/webapp/schemas.py
@@ -260,7 +260,7 @@ class Payment(Schema):
     action = fields.Str(required=True, validate=OneOf(["charge", "auth"]))
     currency = diem_currency_code_field(required=True)
     amount = fields.Int(required=True)
-    expiration = fields.Int(required=False, allow_none=True)
+    expiration = fields.Decimal(required=False, allow_none=True)
 
 
 class PaymentCommand(Schema):
